+++ libs/react/fpkit/src/components/modal/dialog.tsx
@@ -1,16 +1,31 @@
-import FP from "../fp"
-import { ComponentProps } from "../../types"
+import FP from '../fp'
+import { ComponentProps } from '../../types'
+import { Button } from '../buttons/button'
+import React, { MouseEventHandler } from 'react'
 
 export interface DialogProps extends ComponentProps {
-  id: String
-  open: Boolean
-  header: React.ReactNode
-  footer: React.ReactNode
   modalRef: React.RefObject<HTMLDialogElement>
+  closeModal?: (e: React.SyntheticEvent<HTMLDialogElement>) => void
 }
 
-const Dialog = ({children}: DialogProps) => {
-  return <FP as="dialog">{children}</FP>
+const handleCloseModal = (e: React.SyntheticEvent<HTMLDialogElement>) => {
+  if (e.currentTarget === e.target) {
+    e.currentTarget.close()
+  }
 }
 
-export default Dialog
+export const Dialog = ({ id, children, modalRef, ...props }: DialogProps) => {
+  return (
+    <FP
+      as="dialog"
+      id={id}
+      ref={modalRef}
+      onClick={handleCloseModal}
+      {...props}
+    >
+      {children}
+    </FP>
+  )
+}
+
+Dialog.displayName = 'Dialog'