+++ libs/react/fpkit/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.22](https://github.com/shawn-sandy/fpkit/compare/@fpkit/react@0.2.21...@fpkit/react@0.2.22) (2022-11-24)
+
+### Bug Fixes
+
+- 🚨 typo ..props ([3f90dc5](https://github.com/shawn-sandy/fpkit/commit/3f90dc585aba24dca60100d8383fe0e517884a34))
+
+### Features
+
+- 🌟 add support for rest props on modal/dialog ([a7b27ba](https://github.com/shawn-sandy/fpkit/commit/a7b27ba4f2ffa9851b5c9681473c2acaa7842b92))
+
+## [0.2.21](https://github.com/shawn-sandy/fpkit/compare/@fpkit/react@0.2.20...@fpkit/react@0.2.21) (2022-11-24)
+
+### Bug Fixes
+
+- 🐛 tsc/lint fix errors ([411df9a](https://github.com/shawn-sandy/fpkit/commit/411df9a5e86739430f36ff8aa7f072d166dd6b34))
+
+## [0.2.20](https://github.com/shawn-sandy/fpkit/compare/@fpkit/react@0.2.19...@fpkit/react@0.2.20) (2022-11-24)
+
+### Bug Fixes
+
+- 🐛 tsc/lint fix errors ([8f439ea](https://github.com/shawn-sandy/fpkit/commit/8f439eaa357726ec78000cd3b2f889faf108f5ff))
+- 🐛 tsc/lint fix errors ([bfdf0d1](https://github.com/shawn-sandy/fpkit/commit/bfdf0d1b4131bc50e214c607c883ce9171a727e9))
+
+### Features
+
+- 🚧 modal/dialog components ([1f7366b](https://github.com/shawn-sandy/fpkit/commit/1f7366beca0ffddc9033b8f7719cbcdd641424ed))
+- 🚧 stub out the modal component function ([8473097](https://github.com/shawn-sandy/fpkit/commit/84730971e9646fd24fc6af669242890a374c7ded))
+
+## 0.2.19 (2022-11-23)
+
+### Bug Fixes
+
+- 🧯 stop img-error looping ([17deb21](https://github.com/shawn-sandy/fpkit/commit/17deb210ce8819a4fa311c97713f45fc4c21aaf7))
+- storybook sidebar navigation ([f39bcc8](https://github.com/shawn-sandy/fpkit/commit/f39bcc855912e1b9d3100405885935e74e7e7dbc))
+
+### Features
+
+- :recycle: set children pram to be optional type ([d56c336](https://github.com/shawn-sandy/fpkit/commit/d56c336cf5b1519ed1e5c0b74689bab5e6687589))
+- ✨ make span conditional add default span tag ([c4c961a](https://github.com/shawn-sandy/fpkit/commit/c4c961a3e402a6343771a0415ac04328227fc6ac))
+- ✨ set renderstyles to stur by default ([9a466ce](https://github.com/shawn-sandy/fpkit/commit/9a466ce583f87e9bccdd8ba11203b961542faa4a))
+- ❇️ add a card component to fpkit ([e6238f8](https://github.com/shawn-sandy/fpkit/commit/e6238f860ab2458f83c00971d3e098c2e541ec58))
+- 🌟 add display name ([c387ea7](https://github.com/shawn-sandy/fpkit/commit/c387ea7ca29853bb53678629ca2a1e3061ee37da))
+- 🌟 add image onError ([b011d26](https://github.com/shawn-sandy/fpkit/commit/b011d26512cf1ac352a75a37697536628ed92620))
+- 🌟 add img component ([ce857d4](https://github.com/shawn-sandy/fpkit/commit/ce857d4628ea6e0e06ff1a4d59bbed036542d687))
+- 🌟 breadcrumb component ([36685e3](https://github.com/shawn-sandy/fpkit/commit/36685e3b7cda73fd8ed4aa5966ad2eab2aac3062))
+- 🌟 update the breadcrumb listItems ([68ef1b5](https://github.com/shawn-sandy/fpkit/commit/68ef1b5aa9f653d0009ba0992745543d0dddc7fe))
+- 🏗️ add button component ([8bfdf87](https://github.com/shawn-sandy/fpkit/commit/8bfdf87ce5f8e84bd4e3c51efebcf6a8664f4e16))
+- 👆 update disabled button styles ([efad4b2](https://github.com/shawn-sandy/fpkit/commit/efad4b2a61010142974dc5178a079e394e33cec3))
+- 📝 storybooks updates ([86b3662](https://github.com/shawn-sandy/fpkit/commit/86b3662224e7f259ba4452622d6566c42a647963))
+- 🪄 set the placeholder imgWidth to imgWidth ([6843943](https://github.com/shawn-sandy/fpkit/commit/6843943b10dab3e231ec58e7640894225c94da45))
+- 🚧 ad badge component fpkit/fp ([c8b6b7b](https://github.com/shawn-sandy/fpkit/commit/c8b6b7b11951ce8586505a01b2658b79187e3828))
+- 🚧 ad badge component pf fpkit/fp ([f17fd2d](https://github.com/shawn-sandy/fpkit/commit/f17fd2d61cf65cd6150e7cda77debe880cca266a))
+- 🚧 add base breadcrumb component ([de421d6](https://github.com/shawn-sandy/fpkit/commit/de421d658d7c291aaea0bd8fe5ef71ab20519d3c))
+- 🚧 add new component ([fcb8001](https://github.com/shawn-sandy/fpkit/commit/fcb800111d2a92b063c26e8049c51c4d2e503887))
+- 🚧 base dropdown component ([902d945](https://github.com/shawn-sandy/fpkit/commit/902d945fd9100d5bdc266b36edad248b1a7e542f))
+- 🚧 base dropdown component ([aff9e81](https://github.com/shawn-sandy/fpkit/commit/aff9e810866f16ff4d716032c8d30f0703671abc))
+- 🚧 breadcrumb component story ([0a750ee](https://github.com/shawn-sandy/fpkit/commit/0a750eebcf87c0aa7ad8bd9aa594c841d35ad692))
+- 🚧 scaffold dropdown-details component ([3233a88](https://github.com/shawn-sandy/fpkit/commit/3233a8832085b8bcbee1d6fc97b1f1f3294b2a32))
+- 🚧 update card props values ([e852d27](https://github.com/shawn-sandy/fpkit/commit/e852d2715a8597c132ccbe49c7b17615cf51d8db))
+- 🚧 update dropdown component ([80d5006](https://github.com/shawn-sandy/fpkit/commit/80d5006607064cd80b35253ba70f706aa9bc5c91))
+- 🚧 update dropdown component ([b349385](https://github.com/shawn-sandy/fpkit/commit/b349385ec82bc08f5c164720074e69cc2323fa43))
+- add a box component ([b445b5a](https://github.com/shawn-sandy/fpkit/commit/b445b5a5920b5fc892a758089b5ad325cec5bb40))
+- Add content elements to storybook ([07fd928](https://github.com/shawn-sandy/fpkit/commit/07fd928585609e77fa0835ec356525da695fda2a))
+- add onLoad event to the image component ([84cdb1e](https://github.com/shawn-sandy/fpkit/commit/84cdb1e15f153b93edaf16f1cf41210166d8fbad))
+- merge ComponentProps into ImgProps ([1a0ce9d](https://github.com/shawn-sandy/fpkit/commit/1a0ce9d61c785de8588c5cb7766eb241120e35d1))
+
+### Reverts
+
+- 🔥 remove image value (default) ([4358210](https://github.com/shawn-sandy/fpkit/commit/435821040e56ede76c0e8c9d689471834698b508))
+
 ## [0.2.18](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.17...@fpkit/fp@0.2.18) (2022-11-22)
 
 ### Bug Fixes